requirements_conda_list.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
brotlipy=0.7.0=py310he2412df_1003
bzip2=1.0.8=he774522_0
ca-certificates=2021.10.26=haa95532_2
cachetools=4.2.2=pyhd3eb1b0_0
certifi=2021.5.30=py310haa95532_0
cffi=1.15.0=py310hcbf9ad4_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
cryptography=36.0.0=py310ha857299_0
dcmweb=1.0=pypi_0
decorator=5.1.0=pypi_0
dicomweb-client=0.54.2=pypi_0
fire=0.4.0=pypi_0
google-api-core=2.3.2=pypi_0
google-api-python-client=2.32.0=pyhd8ed1ab_0
google-auth=2.3.3=pypi_0
google-auth-httplib2=0.1.0=pyhd8ed1ab_0
google-cloud-core=2.2.1=pypi_0
google-cloud-pubsub=2.9.0=pypi_0
google-cloud-storage=1.43.0=pypi_0
google-crc32c=1.3.0=pypi_0
google-oauth=1.0.1=pypi_0
google-resumable-media=2.1.0=pypi_0
googleapis-common-protos=1.53.0=py310heb1253a_1
grpc-google-iam-v1=0.12.3=pypi_0
grpcio=1.42.0=pypi_0
grpcio-status=1.42.0=pypi_0
httplib2=0.20.2=pyhd8ed1ab_1
hurry-filesize=0.9=pypi_0
idna=3.3=pyhd3eb1b0_0
libcst=0.3.23=pypi_0
libffi=3.4.2=h604cdb4_1
libprotobuf=3.19.1=h7755175_0
libzlib=1.2.11=h8ffe710_1013
mypy-extensions=0.4.3=pypi_0
numpy=1.21.4=pypi_0
oauth2client=4.1.3=pypi_0
openssl=1.1.1l=h2bbff1b_0
pillow=8.4.0=pypi_0
pip=21.2.4=py310haa95532_0
proto-plus=1.19.8=pypi_0
protobuf=3.19.1=py310h8a704f9_1
pyasn1=0.4.8=pyhd3eb1b0_0
pyasn1-modules=0.2.8=py_0
pycparser=2.21=pyhd3eb1b0_0
pydicom=2.2.2=pypi_0
pyopenssl=21.0.0=pyhd3eb1b0_1
pyparsing=3.0.4=pyhd3eb1b0_0
pyqt5=5.15.6=pypi_0
pyqt5-qt5=5.15.2=pypi_0
pyqt5-sip=12.9.0=pypi_0
pysocks=1.7.1=py310h5588dad_4
python=3.10.0=h96c0403_3
python_abi=3.10=2_cp310
pytz=2021.3=pyhd3eb1b0_0
pyyaml=6.0=pypi_0
requests=2.26.0=pyhd3eb1b0_0
retrying=1.3.3=pypi_0
rsa=4.7.2=pyhd3eb1b0_1
setuptools=58.0.4=py310haa95532_0
simplejson=3.17.6=py310he2412df_0
six=1.16.0=pyhd3eb1b0_0
sqlite=3.36.0=h2bbff1b_0
termcolor=1.1.0=pypi_0
tk=8.6.11=h2bbff1b_0
typing-inspect=0.7.1=pypi_0
tzdata=2021e=hda174b7_0
uritemplate=3.0.1=py_0
urllib3=1.26.7=pyhd3eb1b0_0
validators=0.18.2=pypi_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wheel=0.37.0=pyhd3eb1b0_1
win_inet_pton=1.1.0=py310h5588dad_3
wincertstore=0.2=py310haa95532_2
xz=5.2.5=h62dcd97_0
zlib=1.2.11=h8ffe710_1013
